requirements.txt
aiogram~=3.22.0
SQLAlchemy~=2.0.44
python-dotenv~=1.2.1
pydantic-settings~=2.11.0
pydantic~=2.11.10
asyncpg~=0.30.0